+++ utils/leave-room.js
@@ -1,0 +1,5 @@
+const leaveRoom = (userID, chatRoomUsers) => {
+  return chatRoomUsers.filter((user) => user.id != userID)
+}
+
+export { leaveRoom }
